backend/requirements.txt
asgiref==3.11.0
black==23.12.1
click==8.3.1
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
iniconfig==2.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
psycopg==3.3.2
psycopg-binary==3.3.2
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.11.0
pytest==7.4.4
pytest-django==4.7.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.5
